// src/app/shared/models/consulta.class.ts
import { Timestamp } from 'rxjs';
import { Time } from '@angular/common';

export class Consulta {
    uid: string;
    citaUid: string;
    recetaUid: string;
    pacienteUid: string;
    medicoUid: string;
    fecha: string;
    sintomas: string;
    indicaciones: string;
    antecedentes: string;
    medicamento: string;
    total: number;
    saldo: number;

}